+++ amazonorders/orders.py
@@ -12,7 +12,7 @@ from bs4 import Tag
 from amazonorders import util
 from amazonorders.conf import AmazonOrdersConfig
 from amazonorders.entity.order import Order
-from amazonorders.exception import AmazonOrdersError, AmazonOrdersNotFoundError
+from amazonorders.exception import AmazonOrdersError, AmazonOrdersNotFoundError, AmazonOrdersAuthError
 from amazonorders.session import AmazonSession
 
 logger = logging.getLogger(__name__)
@@ -56,6 +56,10 @@ class AmazonOrders:
 
         order_details_response = self.amazon_session.get(
             f"{self.config.constants.ORDER_DETAILS_URL}?orderID={order_id}")
+        if order_details_response.response.url.startswith(self.config.constants.SIGN_IN_URL):
+            raise AmazonOrdersAuthError("Amazon redirected to login. Call AmazonSession.login() to "
+                                        "reauthenticate first.")
+
         if not order_details_response.response.url.startswith(self.config.constants.ORDER_DETAILS_URL):
             raise AmazonOrdersNotFoundError(f"Amazon redirected, which likely means Order {order_id} was not found.")
 
@@ -107,6 +111,9 @@ class AmazonOrders:
 
         while next_page:
             page_response_parsed = self.amazon_session.get(next_page).parsed
+            if page_response_parsed.response.url.startswith(self.config.constants.SIGN_IN_URL):
+                raise AmazonOrdersAuthError("Amazon redirected to login. Call AmazonSession.login() to "
+                                            "reauthenticate first.")
 
             for order_tag in util.select(page_response_parsed,
                                          self.config.selectors.ORDER_HISTORY_ENTITY_SELECTOR):
@@ -146,6 +153,9 @@ class AmazonOrders:
                 # TODO: be on the lookout for if this causes rate limit issues with Amazon, or races with the
                 #  URL connection pool. If so, we may need to implement some retry logic here.
                 order_details_response_parsed = self.amazon_session.get(order.order_details_link).parsed
+                if order_details_response_parsed.response.url.startswith(self.config.constants.SIGN_IN_URL):
+                    raise AmazonOrdersAuthError("Amazon redirected to login. Call AmazonSession.login() to "
+                                                "reauthenticate first.")
                 order_details_tag = util.select_one(order_details_response_parsed,
                                                     self.config.selectors.ORDER_DETAILS_ENTITY_SELECTOR)
                 order = self.config.order_cls(order_details_tag, self.config, full_details=True, clone=order,